+++ targets/user/ip-stateless.cpp
@@ -62,7 +62,7 @@ int dst_ip_matching(uint32_t dst_addr) {
 
 int main() {
     uint32_t max_intfs = num_interfaces();
-    uint32_t intf = 0;
+    uint8_t intf = 0;
     Packet ingress_pkt;
     memset(&ingress_pkt, 0, sizeof(ingress_pkt));
     info("Total interfaces: " + std::to_string(max_intfs));
